core_tests: add --filter to select which tests to run /monero#4110
