+++ src/common/util.h
@@ -231,4 +231,6 @@ namespace tools
   bool is_hdd(const char *path);
 
   boost::optional<std::pair<uint32_t, uint32_t>> parse_subaddress_lookahead(const std::string& str);
+
+  std::string glob_to_regex(const std::string &val);
 }
