+++ tests/performance_tests/main.cpp
@@ -55,25 +55,6 @@
 
 namespace po = boost::program_options;
 
-std::string glob_to_regex(const std::string &val)
-{
-  std::string newval;
-
-  bool escape = false;
-  for (char c: val)
-  {
-    if (c == '*')
-      newval += escape ? "*" : ".*";
-    else if (c == '?')
-      newval += escape ? "?" : ".";
-    else if (c == '\\')
-      newval += '\\', escape = !escape;
-    else
-      newval += c;
-  }
-  return newval;
-}
-
 int main(int argc, char** argv)
 {
   TRY_ENTRY();
@@ -98,7 +79,7 @@ int main(int argc, char** argv)
   if (!r)
     return 1;
 
-  const std::string filter = glob_to_regex(command_line::get_arg(vm, arg_filter));
+  const std::string filter = tools::glob_to_regex(command_line::get_arg(vm, arg_filter));
 
   performance_timer timer;
   timer.start();
